Cast pixel status to uint8

+++ lstchain/scripts/lstchain_r0g_to_r0v.py
@@ -161,7 +161,7 @@ def main():
                 new_status = np.where(ordered_pix_mask,
                                       pixel_status | PixelStatus.DVR_STATUS_0,
                                       pixel_status & ~(PixelStatus.DVR_STATUS_0 | PixelStatus.DVR_STATUS_1))
-                event.pixel_status.data = new_status.tobytes()
+                event.pixel_status.data = np.uint8(new_status).tobytes()
 
                 stream.write_message(event)
